+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ### Fixed
 
+- `Polyline.Split(List<Vector3> point)`
+- `Polygon.Split(List<Vector3> point)`
+- `Polygon.TrimmedTo(List<Polygon> polygons)`
+- `Vector3.>=`
+- `Vector3.<=`
+- `Plane.Intersects(Plane a, Plane b)`
+
 ## 0.9.1
 
 ### Added